refactor: reorganizar estrutura de autenticação
